requirements.txt
absl-py==2.1.0
accelerate==0.24.1
aiohttp==3.9.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==23.2.0
blip-ci==0.0.5
certifi==2023.11.17
charset-normalizer==3.3.2
clip-interrogator==0.5.4
datasets==2.18.0
diffusers==0.27.2
dill==0.3.8
fairscale==0.4.4
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.10.0
ftfy==6.2.0
grpcio==1.62.1
huggingface-hub==0.22.2
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.4.2
Markdown==3.6
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
open-clip-torch==2.24.0
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.1
Pillow==10.1.0
pip==23.3.1
protobuf==5.26.1
psutil==5.9.6
pyarrow==15.0.2
pyarrow-hotfix==0.6
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
scikit-learn==1.4.2
scipy==1.13.0
sentence-transformers==2.7.0
sentencepiece==0.1.99
setuptools==68.2.2
six==1.16.0
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
timm==0.4.12
tokenizers==0.15.0
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
transformers==4.35.2
triton==2.1.0
typing_extensions==4.8.0
tzdata==2024.1
urllib3==2.1.0
wcwidth==0.2.13
Werkzeug==3.0.2
wheel==0.41.2
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
